requirements.txt
jsonrpc
bitcoin-python
pymongo
mailer
flask
python-bitcoinrpc
